Edge heading quicknav: re-implement using move by paragraph and getAttributeValue rather than findAttribute.
Using findAttribute was never totally accurate in all situations, and its speed was based on the size of the document rather than how far the next heading was. In most cases moving by heading is now faster. In a few cases it may have become slower, though it is now more in line with Narrator.